A better battery segment
Uses the UPower service on linux to display the actual time remaining
before the battery finishes charging/discharging. Also hides itself
automatically if there is no battery or if the battery is fully charged
and on AC power. I haven't bothered with adding the segment to any
themes, as I have a custom theme setup in any case. If someone merges
it, it should probably be put into the tmux and wm themes at least
(replacing the current battery segment).